# FJT Recursion Change