--- a/psi4/src/psi4/libmints/eri.cc
+++ b/psi4/src/psi4/libmints/eri.cc
@@ -41,11 +41,11 @@ ERI::ERI(const IntegralFactory *integral, int deriv, bool use_shell_pairs)
     : TwoElectronInt(integral, deriv, use_shell_pairs)
 {
     // The +1 is needed for derivatives to work.
-    fjt_ = new Taylor_Fjt(basis1()->max_am() +
-                          basis2()->max_am() +
-                          basis3()->max_am() +
-                          basis4()->max_am() +
-                          deriv_+1, 1e-15);
+    fjt_ = new Split_Fjt(basis1()->max_am() +
+                         basis2()->max_am() +
+                         basis3()->max_am() +
+                         basis4()->max_am() +
+                         deriv_+1);
 }
 
 ERI::~ERI()