@bennybp 2017-01-30
I have a good setup for vtune on some clusters. Do you have any inputs/scripts I can test?

Tangentially related, I'm getting a better look at the overall thread scaling. In density-fitted HF, the infamous dcopy is a very prominent problem - it's memory bound, so it takes only a few cores for the memory bandwidth to be saturated.

If there's any other specific types of calculations you'd like to see, let me know.