@dgasmith 2017-01-30
@bennybp I used this [gist](https://gist.github.com/dgasmith/876687198a62d1af417ab7f0622b48fb) to time the integrals. 

If you have the output for a large `ao_potential` I would be curious to see what is taking time in `compute_shell` for 4+ threads. Im guessing its an IOPS thing as well.

Well, at least we should have the dcopy bottleneck removed in Psi4 1.2. Before then we could write you a dense DFJK if needed.