@bennybp 2017-01-30
Some bottlenecks that should be easy to fix:

1.) osrecur.cc:998 Allocates (a tiny bit) of space on the heap. Can be fixed with an array on the stack
2.) The calculate_f function
3.) It might help to allocate just a contiguous array of memory in "init_box", etc.

Just changing the allocation of F from heap to stack improved the timings by 10-20% for overlap, up to 50% for ao_potential in my tests (would be interesting if you can verify that).

Thread scaling is better, but could still be improved I think. I may need to get timings for bigger tests