@jturney 2017-01-31
Yeah, I think the allocation at osrecur.cc:998 was a "let's get a version working and I'll refactor later" and later never occurred.