@bennybp 2017-01-31
Mind if I hack around with the FJT/Boys function stuff a bit? It may not be a straightforward replacement, but I think it can be simplified quite a bit