@jturney 2017-01-31
Go for it!

> On Jan 31, 2017, at 10:55 AM, Ben Pritchard <notifications@github.com> wrote:
> 
> Mind if I hack around with the FJT/Boys function stuff a bit? It may not be a straightforward replacement, but I think it can be simplified quite a bit
> 
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub, or mute the thread.
>