@bennybp 2017-01-31
I've got a new version of the fjt code ready to test, however it's based off of master. There is a conflict with this branch and master (export_mints.cc). Could someone more familiar with the changes rebase/merge master into this?