@dgasmith 2017-01-31
@bennybp Yea, that one was pretty subtle. Rebased up to master, please do a `reset --hard`.