@dgasmith 2017-02-03
Full tests pass except DCFT5/6 which attempts to compare the energy to 10 decimal places. On that note it looks like the SCF/CCSD energy difference is about 5.e-10 or so. Not too surprising since the one-electron integrals are slightly different. Looks like the integrals for s/p are about the same, but for d+ its considerably faster.

Might try to rig up a test where we check the norm between the current and previous versions matrix elements for large basis sets and heavy atoms.

```
# 6 core update.
   nbf  ntrial  ao_overlap  ao_kinetic  ao_potential  so_overlap  so_kinetic  so_potential
0   28    1275    0.001758    0.000034      0.001248    0.001832    0.000177      0.001438
1   98     104    0.002072    0.000253      0.004509    0.002503    0.000808      0.005072
2  168      35    0.002780    0.000696      0.017807    0.003713    0.001938      0.019500
3  238      17    0.003789    0.001310      0.044333    0.005827    0.003503      0.047694
4  308      10    0.005321    0.002705      0.093579    0.009500    0.006423      0.096568
5  378       6    0.007993    0.004080      0.178422    0.012842    0.009268      0.173306
6  448       4    0.011465    0.006115      0.276808    0.017397    0.012930      0.282091
7  518       3    0.015315    0.007962      0.426757    0.023739    0.018704      0.436435
```