@dgasmith 2017-02-06
Taking the Krypton dimer at 3 angstroms with the def2-QZVP basis we get:

```
SCF Energy difference: 7.1e-10 [hartree]
AO potenital norm :    2.1e-8
AO max diff:           5.1e-9

AO ERI norm:           4.6e-8
AO ERI max diff:       2.0e-9
```

Kr is nice and heavy with up to G functions. We could test something like aug-cc-pVDZ-RI for H and I functions.