@dgasmith 2017-10-24
Rebased this for discussion during the Psi4 meeting.