@dgasmith 2018-08-02
@andysim I think we already merged the threading changes.