@loriab 2018-09-03
From Andy: current status is that the 1e-10 errors are suspect. Want to use Ben's MIRP https://github.com/MolSSI/MIRP to gauge what's right and see if they're acceptable. Good miniproject.